refactor(scripts): #72 2.2-3 — extract TE planner+runner to scripts/validate/te.py
Move TE constants (TE_PARSER_READS, TE_UPDATES_POINTER_READS,
TE_ENUM_FIELDS), the TE planner, diff_te_row, try_parse, run_probe,
and resolve_dynamic_ids into scripts/validate/te.py (222 LOC). The
validator drops 3,930 → 3,732 LOC.

Cross-cutting helper bump in `_shared.py`: the ISO date utilities
(`today_iso`, `days_ago_iso`, `days_ahead_iso`) move there too — they
were previously script-private with leading-underscore names but are
needed by both TE and EODHD planners. Promoting them to public
`_shared` API avoids a back-import from the new TE module to the
parent script. EODHD's `_days_ahead_iso` (its only local copy of the
pattern) moves with them.

Subtraction pass: removed two now-dead imports — `MUTABLE_FIELDS`
(unused since the validator was first written) and `urllib.parse.quote`
(used only by the TE planner that just moved).

Tests: 9 scaffold tests load the validator unchanged at 135 passed;
broader `-k "scaffold or validate or scripts"` subset at 695 passed.
Dry-runs for `--provider te` and `--provider eodhd` both succeed
(confirmed by codex review).

Refs #72.
